pyproject.toml
[project]
name = "plant-doctor"
version = "0.0.2"
description = "A Google Agent for plant disease diagnosis."
authors = [{ name = "Nancy Goyal", email = "nancy@example.com"  }]
classifiers = [
    "Programming Language :: Python",    
    "Programming Language :: Python :: 3.10",
]
keywords = ['agent-packs', 'adk', 'agent']
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pytest-asyncio",
    "tensorflow",
    "numpy",
    "google-adk",
    "google-generativeai",
    "langchain-chroma",
    "langchain-google-genai",
    "langchain_openai",
    "pypdf",
    "openai",
    "chromadb",
    "pyyaml",
    "uvicorn",
    "Pillow",
    "litellm",
]

[build-system]
requires = ["setuptools >= 40.9.0",]
build-backend = "setuptools.build_meta"


[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipython>=8.36.0",
    "pytest>=8.3.5",
    "pre-commit>=4.2.0",
    "ruff>=0.11.5",
    "mypy>=1.16.0",
    "deptry>=0.23.0",
    "pytest-cov>=4.0.0",
    "types-pyyaml>=6.0.12.20250809",
]

[tool.agent_packs]
template = "adk"

[tool.mypy]
files = ["src"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true
ignore_missing_imports = true

[tool.basedpyright]
reportAny = false
reportExplicitAny = false
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownVariableType = false

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
fix = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["src"]
omit = ["server.py"]

[tool.deptry.per_rule_ignores]
DEP002 = ["uvicorn"]